move hangout file json parsing out of main
